services/vision-pipeline/pyproject.toml
[project]
name = "precept-vision-pipeline"
version = "0.1.0"
description = "Local-first vision pipeline service for Precept."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic-settings>=2.4.0",
  "numpy>=1.26.0",
  "Pillow>=10.4.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
